feat(mapview): Added functionality to render mower position and path
